requirements.txt
torch
torchvision
matplotlib
numpy
streamlit
Pillow